// src/types/menu-items.enum.ts
export enum MenuItem {
    Home = "home",
    BrandManagement = "brandManagement",
    DigitalMarketing = "digitalMarketing",
    MarketingResearch = "marketingResearch",
    MediaPlanning = "mediaPlanning",
    MarketingStrategy = "marketingStrategy",
    ProposalAndPitchDevelopment = "proposalAndPitchDevelopment",
    ExperientialMarketing = "experientialMarketing"
}